refactor: split api calls from hook files
