+++ frontend/hooks/useBoard.tsx
@@ -1,18 +1,7 @@
 import { useRouter } from "next/router";
 import { useMutation } from "react-query";
 import { BoardType, UseBoardType } from "../types/boardTypes";
-import { Nullable } from "../types/commonTypes";
-import fetchData from "../utils/fetchData";
-
-const API_URL = process.env.NEXT_PUBLIC_BACKEND_URL;
-
-export const postBoard = async (newBoard: BoardType): Promise<BoardType> => {
-  return fetchData(`${API_URL}/boards.json`, "POST", JSON.stringify(newBoard));
-};
-
-export const getBoard = async (id: Nullable<string>): Promise<BoardType> => {
-  return fetchData(`${API_URL}/boards/${id}.json`, "GET", null);
-};
+import { postBoard } from "../api/boardService";
 
 const useBoard = (): UseBoardType => {
   const router = useRouter();
